+++ src/KodiRPC/Responses/Types/Video/Resume.cs
@@ -5,9 +5,9 @@ namespace KodiRPC.Responses.Types.Video
     public class Resume
     {
         [JsonProperty(PropertyName = "position")]
-        public int Position { get; set; }
+        public double Position { get; set; }
 
         [JsonProperty(PropertyName = "total")]
-        public int Total { get; set; }
+        public double Total { get; set; }
     }
 }
